requirements.txt
PyQt6==6.6.1
python-dotenv==1.0.0
mysql-connector-python==8.2.0
